requirements.txt
streamlit>=1.42.0
st-supabase-connection>=2.0.1
groq>=0.18.0
pdfplumber>=0.11.5
filetype>=1.2.0
gotrue
langchain
langchain-community
langchain-huggingface
faiss-cpu
sentence-transformers
langchain-text-splitters
